+++ verification/stdlib/concurrency_runtime_m4_process_traceability.md
@@ -2,7 +2,7 @@
 
 Milestone: `milestone_concurrency_runtime_4`
 
-Status: In progress; sync process foundation merged in PR #2331, sync child wait merged in PR #2334, timeout status evidence merged in PR #2336, sync child kill support merged in PR #2337, Unix signal-status evidence merged in PR #2341, legacy subprocess intrinsic cleanup merged in PR #2344, async run/output loopback merged in PR #2345, stdin append semantics merged in PR #2348, method-form blocking diagnostics merged in PR #2350, sync stdout/stderr pipe readers merged in PR #2352, async run timeout merged in PR #2354, sync stdin pipe writer support merged in PR #2357, stdin guardrails merged in PR #2359, async output timeout merged in PR #2362, async stdin-byte communicate merged in PR #2365, sync terminate merged in PR #2367, async spawn/wait merged in PR #2369, and method-form async child kill/terminate merged in PR #2372.
+Status: In progress; sync process foundation merged in PR #2331, sync child wait merged in PR #2334, timeout status evidence merged in PR #2336, sync child kill support merged in PR #2337, Unix signal-status evidence merged in PR #2341, legacy subprocess intrinsic cleanup merged in PR #2344, async run/output loopback merged in PR #2345, stdin append semantics merged in PR #2348, method-form blocking diagnostics merged in PR #2350, sync stdout/stderr pipe readers merged in PR #2352, async run timeout merged in PR #2354, sync stdin pipe writer support merged in PR #2357, stdin guardrails merged in PR #2359, async output timeout merged in PR #2362, async stdin-byte communicate merged in PR #2365, sync terminate merged in PR #2367, async spawn/wait merged in PR #2369, method-form async child kill/terminate merged in PR #2372, and async process runtime decomposition is in the current implementation wave.
 
 ## Production Surface Traceability
